Move NeutronSource class around
